LayoutTests/fast/dom/Geolocation/reentrant-error-expected.txt

Tests that reentrant calls to Geolocation methods from the error callback are OK.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true
PASS error.code is mockCode
PASS error.message is mockMessage
PASS error.UNKNOWN_ERROR is 0
PASS error.PERMISSION_DENIED is 1
PASS error.POSITION_UNAVAILABLE is 2
PASS error.TIMEOUT is 3

PASS error.code is mockCode
PASS error.message is mockMessage
PASS error.UNKNOWN_ERROR is 0
PASS error.PERMISSION_DENIED is 1
PASS error.POSITION_UNAVAILABLE is 2
PASS error.TIMEOUT is 3

TEST COMPLETE
